requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.4
pandas==0.25.0
pkg-resources==0.0.0
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.6
Werkzeug==0.15.5
